requirements.txt
PyYAML==6.0.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
python-dateutil==2.8.2
